Updates mimic joint list on URDF swap
This change fixes the issue preventing mimic joints Transforms (TF's)
from be updated with the URDF swapping schema. Previously, the MimicMap
was set at the start time of the robot_state_publisher and could never
be updated, even when a new mimic joint was added via URDFConfiguration
message. Now, the RobotStatePublisher has complete control over the
MimicMap and the JointStatePublisher needs to request an update on every
callbackJointStates() invocation. The MimicMap is read-mutex locked for
the callbackJointStates() calls, and exlusive write-mutex locked when
it needs to updated on a URDF Swap. If the read-mutex fails, the a
ROS WARN is issued and the mimic joint transforms are not updated.